+++ src/main/java/com/emaciejko/dao/CustomerDAO.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.emaciejko.domain.Customer;
 
 @Repository
-public interface CustomerDAO extends CrudRepository<Customer, Long> {  
+public interface CustomerDAO extends CrudRepository<Customer, Integer> {
